move useChangePassword()
