+++ projects/graphql-client/src/index.ts
@@ -34,3 +34,4 @@ export * from "./remote/mutations/useDeleteShare"
 export * from "./remote/mutations/useDeletePlaylist"
 export * from "./remote/mutations/useCreateShare"
 export * from "./remote/mutations/useCreatePlaylist"
+export * from "./remote/mutations/useChangePassword"
